Код разбит на исполняемый ф., функции и константы
